Rename existing sprint module to general_labels
