+++ test/pages/print_labels/source_labels.spec.js
@@ -3,12 +3,12 @@ import { BootstrapVue } from 'bootstrap-vue'
 import { mount, createLocalVue } from '@vue/test-utils'
 import SourcePlates from '@/pages/print_labels/source_plates'
 import statuses from '@/modules/statuses'
-import Sprint from '@/modules/sprint'
+import Sprint from '@/modules/sprint_general_labels'
 import CSV from '@/modules/csv'
 import config from '@/nuxt.config'
 import barcodes from '@/test/data/barcodes'
 
-jest.mock('@/modules/sprint')
+jest.mock('@/modules/sprint_general_labels')
 jest.mock('@/modules/csv')
 
 const localVue = createLocalVue()
